Split off functions and whenthen
