+++ py-polars/polars/eager/__init__.py
@@ -1,3 +1,5 @@
 # flake8: noqa
 from .frame import *
 from .series import *
+
+__all__ = frame.__all__ + series.__all__
