+++ py-polars/polars/lazy/__init__.py
@@ -1,3 +1,7 @@
 # flake8: noqa
 from .expr import *
 from .frame import *
+from .functions import *
+from .whenthen import *
+
+__all__ = expr.__all__ + frame.__all__ + functions.__all__ + whenthen.__all__
